Add Windows support and refactor API
The API was changed to use opaque data types to aid porting to different platforms. Support
for Windows platforms was added and common functionality was separated into groups
with similar concerns.

The README was improved and an example program was added.

Compilation was tested with MingW32, GCC and Visual Studio 12. Functionality was tested
on Linux and Windows 10 platforms.